+++ concordium-sdk/src/main/java/com/concordium/sdk/transactions/TransactionFactory.java
@@ -109,6 +109,12 @@ public class TransactionFactory {
         return ConfigureBakerTransaction.builder();
     }
 
+    /**
+     * Creates a new {@link ConfigureBakerTransaction.ConfigureBakerTransactionBuilder} for
+     * creating a {@link ConfigureBakerTransaction} to remove Baker
+     *
+     * @return the builder for a {@link ConfigureBakerTransaction}
+     */
     public static ConfigureBakerTransaction.ConfigureBakerTransactionBuilder newRemoveBaker() {
         val payload = ConfigureBakerPayload.builder()
                 .capital(CCDAmount.fromMicro(0))
@@ -116,13 +122,25 @@ public class TransactionFactory {
         return ConfigureBakerTransaction.builder().payload(payload);
     }
 
-    public static ConfigureBakerTransaction.ConfigureBakerTransactionBuilder newUpdateBakerStake(long stake) {
+    /**
+     * Creates a new {@link ConfigureBakerTransaction.ConfigureBakerTransactionBuilder} for
+     * creating a {@link ConfigureBakerTransaction} to update Baker Stake
+     *
+     * @return the builder for a {@link ConfigureBakerTransaction}
+     */
+    public static ConfigureBakerTransaction.ConfigureBakerTransactionBuilder newUpdateBakerStake(CCDAmount stake) {
         val payload = ConfigureBakerPayload.builder()
-                .capital(CCDAmount.fromMicro(stake))
+                .capital(stake)
                 .build();
         return ConfigureBakerTransaction.builder().payload(payload);
     }
 
+    /**
+     * Creates a new {@link ConfigureBakerTransaction.ConfigureBakerTransactionBuilder} for
+     * creating a {@link ConfigureBakerTransaction} to update restakeEarnings
+     *
+     * @return the builder for a {@link ConfigureBakerTransaction}
+     */
     public static ConfigureBakerTransaction.ConfigureBakerTransactionBuilder newUpdateBakerRestakeEarnings(boolean restakeEarnings) {
         val payload = ConfigureBakerPayload.builder()
                 .restakeEarnings(restakeEarnings)
@@ -130,9 +148,15 @@ public class TransactionFactory {
         return ConfigureBakerTransaction.builder().payload(payload);
     }
 
+    /**
+     * Creates a new {@link ConfigureBakerTransaction.ConfigureBakerTransactionBuilder} for
+     * creating a {@link ConfigureBakerTransaction} to update baker keys
+     *
+     * @return the builder for a {@link ConfigureBakerTransaction}
+     */
     public static ConfigureBakerTransaction.ConfigureBakerTransactionBuilder newUpdateBakerKeys(AccountAddress accountAddress) {
         val payload = ConfigureBakerPayload.builder()
-                .keysWithProofs(ConfigureBakerKeysPayload.newBakerKeysWithPayload(accountAddress))
+                .keysWithProofs(ConfigureBakerKeysPayload.getNewConfigureBakerKeysPayload(accountAddress))
                 .build();
         return ConfigureBakerTransaction.builder().payload(payload);
     }